Answers are correctly mapped to the question
