+++ client/components/qa_components/QuestionList.jsx
@@ -10,68 +10,65 @@ import token from './config/config.js';
 const QuestionList = (props) => {
   // local State here
   const [currentQuestion, setCurrentQuestion] = useState([]);
-  const [allAnswers, setAllAnswers] = useState([]); // all answers for ONE question
-  const [displayedAnswers, setDisplayedAnswers] = useState([]);
-  const [allAnswersCount, setAllAnswersCount] = useState([]);
-  const [displayedAnswersCount, setDisplayedAnswersCount] = useState([]);
 
-  // get answers
-  const answerPageCount = useRef(1);
-  const getAnswers = (questionId) => {
-    const config = {
-      headers: { Authorization: token },
-    }
-    const queryParam = questionId;
-    axios.get(`https://app-hrsei-api.herokuapp.com/api/fec2/hr-atx/qa/questions/${queryParam}/answers`, config)
-      .then((results) => {
-        // NEED TO SORT THE DATA BEFORE SETTING STATE
-        setAllAnswers(allAnswers => {
-          return [...allAnswers, ...results.data.results]
-        })
-        console.log(results.data.results);
-      })
-      .catch((err) => {
-        console.error('Error: ', err);
-      })
-  }
 
-  const getDisplayedAnswers = (questionId) => {
-    const config = {
-      headers: { Authorization: token },
-      params: {
-        page: 1,
-        count: 2
-      }
-    }
-    const queryParam = questionId;
-    axios.get(`https://app-hrsei-api.herokuapp.com/api/fec2/hr-atx/qa/questions/${queryParam}/answers`, config)
-      .then((results) => {
-        // NEED TO SORT THE DATA BEFORE SETTING STATE
-        setDisplayedAnswers(displayedAnswers => {
-          return [...displayedAnswers, ...results.data.results]
-        })
-      })
-      .catch((err) => {
-        console.error('Error: ', err);
-      })
+  // // get answers
+  // const answerPageCount = useRef(1);
+  // const getAnswers = (questionId) => {
+  //   const config = {
+  //     headers: { Authorization: token },
+  //   }
+  //   const queryParam = questionId;
+  //   axios.get(`https://app-hrsei-api.herokuapp.com/api/fec2/hr-atx/qa/questions/${queryParam}/answers`, config)
+  //     .then((results) => {
+  //       // NEED TO SORT THE DATA BEFORE SETTING STATE
+  //       setAllAnswers(allAnswers => {
+  //         return [...allAnswers, ...results.data.results]
+  //       })
+  //       console.log(results.data.results);
+  //     })
+  //     .catch((err) => {
+  //       console.error('Error: ', err);
+  //     })
+  // }
 
-  };
+  // const getDisplayedAnswers = (questionId) => {
+  //   const config = {
+  //     headers: { Authorization: token },
+  //     params: {
+  //       page: 1,
+  //       count: 2
+  //     }
+  //   }
+  //   const queryParam = questionId;
+  //   axios.get(`https://app-hrsei-api.herokuapp.com/api/fec2/hr-atx/qa/questions/${queryParam}/answers`, config)
+  //     .then((results) => {
+  //       // NEED TO SORT THE DATA BEFORE SETTING STATE
+  //       setDisplayedAnswers(displayedAnswers => {
+  //         return [...displayedAnswers, ...results.data.results]
+  //       })
+  //     })
+  //     .catch((err) => {
+  //       console.error('Error: ', err);
+  //     })
+
+  // };
 
 
 
-  // load answers
-  const moreAnswers = (e) => {
-    console.log('clicked')
-    console.log(props.displayedQs);
-    // answerPageCount.current++;
-    // getAnswers(answerPageCount.current);
-  }
+  // // load answers
+  // const moreAnswers = (e) => {
+  //   console.log('clicked')
+  //   console.log(props.displayedQs);
+  //   // answerPageCount.current++;
+  //   // getAnswers(answerPageCount.current);
+  // }
 
 
-  // show answers on page load
-  useEffect(() => {
-    // getAnswers(props.productId);
-  }, []);
+  // // show answers on page load
+  // useEffect(() => {
+  //   getAnswers(props.productId);
+  // }, []);
 
 
   // Map over each question and create a QuestionCard for each question
